+++ src/components/Containers.tsx
@@ -48,7 +48,7 @@ import { SwitchField } from './SwitchField';
 import { CheckboxField } from './CheckboxField';
 import { SelectField } from './SelectField';
 import { TextAreaField } from './TextAreaField';
-import { JsonField } from './JsonField';
+import { CodeField } from './CodeField';
 import { ColorField } from './ColorField';
 import DateField from './DateField';
 import FileField from './FileField';
@@ -72,8 +72,8 @@ export const renderField = (
       Component = TextAreaField;
       break;
 
-    case 'json':
-      Component = JsonField;
+    case 'code':
+      Component = CodeField;
       break;
 
     case 'number':
@@ -212,7 +212,7 @@ const emptyFields = {
   object: {},
   checkbox: [],
   select: {},
-  json: {},
+  code: '',
   date: '',
   custom: {},
   color: '',
@@ -229,7 +229,7 @@ export const ArrayField: FC<FieldProps<ArrayFieldSchema>> = ({
   const {
     label,
     isRequired,
-    isCollapsable,
+    isCollapsible,
     defaultIsOpen,
     itemField,
     helperText,
@@ -251,7 +251,7 @@ export const ArrayField: FC<FieldProps<ArrayFieldSchema>> = ({
   const { fields, append, remove, move } = useFieldArray({ name, control });
 
   const { isOpen, onOpen, onToggle } = useDisclosure({
-    defaultIsOpen: !isCollapsable || defaultIsOpen,
+    defaultIsOpen: !isCollapsible || defaultIsOpen,
   });
 
   const arrayStyles = useStyles<ArrayFieldStyles>('arrayField', styles);
@@ -312,7 +312,7 @@ export const ArrayField: FC<FieldProps<ArrayFieldSchema>> = ({
                 disabled={isReadOnly || disabled || readOnly}
                 {...arrayStyles.clearButton}
               />
-              {isCollapsable && (
+              {isCollapsible && (
                 <IconButton
                   icon={isOpen ? <ViewOffIcon /> : <ViewIcon />}
                   aria-label={isOpen ? 'Hide items' : 'Show items'}
@@ -454,7 +454,7 @@ export const ObjectField: FC<FieldProps<ObjectFieldSchema>> = ({
 }) => {
   const {
     label,
-    isCollapsable,
+    isCollapsible,
     isRequired,
     helperText,
     shouldDisplay,
@@ -504,7 +504,7 @@ export const ObjectField: FC<FieldProps<ObjectFieldSchema>> = ({
               )}
             </FormLabel>
           )}
-          {isCollapsable && (
+          {isCollapsible && (
             <IconButton
               icon={isOpen ? <ViewOffIcon /> : <ViewIcon />}
               aria-label={isOpen ? 'Hide items' : 'Show items'}